requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-multiselectfield==0.1.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
whitenoise==6.6.0
